Add toy authentication
App displays a login page if user is not logged in, only after "signing in" can the user see the rest of the app.

Currently backend sends the same token no matter what was in the POST request.

Will handle actual backend authentication code in a different branch most likely.